2-backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
motor>=3.3.0
pydantic>=2.0.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
email-validator>=2.1.0
pyjwt>=2.8.0
python-multipart>=0.0.9
httpx>=0.27.0
numpy
onnxruntime
opencv-python-headless
mediapipe
ollama
ultralytics>=8.1.0
torch>=2.0.0
torchvision>=0.15.0
pymupdf
pytesseract
pdf2image
Pillow